refactor(web): extract the Team creation form so onboarding can mount it
PR-A2 (#49) adds an onboarding flow whose first step is a Team placeholder,
commented "Creating the first Team needs POST /api/v1/teams, which PR-A1 adds".
Rather than have that step grow a second Team creation form, the form moves into
its own module with the page reduced to chrome around it.

The caller decides what happens after creation, so /teams/new can select the new
Team and go to Agents while an onboarding step can advance its own progress
instead. Field derivation, the dropped-character warning and the error surface
stay in one place either way.
